server FDS support looks good

+++ gui/ajax.php
@@ -56,7 +56,7 @@ function ajaxChangeActiveScenario() { #{{{
 function ajaxLaunchSimulation() { #{{{
 	$aamks=getenv("AAMKS_PATH");
 	$working_home=$_SESSION['main']['working_home'];
-	if(!is_file("$working_home/cad.json")) { 
+	if(!is_file("$working_home/cad.json") && !is_file("$working_home/cadfds.json")) { 
 		echo json_encode(array("msg"=>"You need to draw and save the project in <a class=blink href=/aamks/apainter>Apainter</a> first", "err"=>1, "data"=>''));
 		return;
 	}
